refactor(Notice) : 인프라 계층 패키지 구조 개선
- crawler라는 이름이 애플리케이션계층의 crawling과 다를 이유가 없어서, 일관성 차원에서 수정
- ai라는 이름보다 "요약"이라는 책임을 갖고있어서 summarize로 이름 변경
- 기존 format패키지내의 클래스들은 크롤링된 데이터 모델을 정의한것에 가깝기때문에 model로 네이밍 개선
- 기존에 참조하던 클래스들의 import 경로도 수정
