+++ backend/src/main/java/org/cathori/backend/notice/infra/crawling/CrawlingScheduler.java
@@ -1,10 +1,10 @@
-package org.cathori.backend.notice.infra.crawler;
+package org.cathori.backend.notice.infra.crawling;
 
 import java.util.List;
 
 import org.cathori.backend.notice.application.crawling.CrawledNotice;
 import org.cathori.backend.notice.application.crawling.NoticeService;
-import org.cathori.backend.notice.infra.crawler.source.DepartmentSource;
+import org.cathori.backend.notice.infra.crawling.source.DepartmentSource;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Scheduled;
