+++ backend/src/main/java/org/cathori/backend/notice/infra/crawling/model/NoticeRow.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.notice.infra.crawler.format;
+package org.cathori.backend.notice.infra.crawling.model;
 
 /***
  * 공지 목록 페이지에서
