+++ backend/src/main/java/org/cathori/backend/notice/infra/summarization/AiSummaryResult.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.notice.infra.ai;
+package org.cathori.backend.notice.infra.summarization;
 
 import java.util.List;
