+++ backend/src/main/java/org/cathori/backend/notice/infra/summarization/RetryScheduler.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.notice.infra;
+package org.cathori.backend.notice.infra.summarization;
 
 import lombok.RequiredArgsConstructor;
 import org.cathori.backend.notice.application.crawling.NoticeService;
