+++ backend/src/main/java/org/cathori/backend/notification/application/push/PushNotificationService.java
@@ -9,7 +9,7 @@ import java.util.stream.Collectors;
 
 import org.cathori.backend.notification.domain.AlertHistory;
 import org.cathori.backend.notification.domain.AlertHistoryRepository;
-import org.cathori.backend.notice.infra.crawler.source.DepartmentSource;
+import org.cathori.backend.notice.infra.crawling.source.DepartmentSource;
 import org.cathori.backend.notice.model.Notice;
 import org.cathori.backend.notice.model.NoticeRepository;
 import org.cathori.backend.user.domain.User;
