+++ backend/src/test/java/org/cathori/backend/notification/application/AlertServiceTest.java
@@ -5,7 +5,7 @@ import org.cathori.backend.notification.application.push.*;
 import org.cathori.backend.notification.domain.AlertHistory;
 import org.cathori.backend.notification.infra.AlertHistoryJpaRepository;
 import org.cathori.backend.notice.application.crawling.CrawledNotice;
-import org.cathori.backend.notice.infra.crawler.source.DepartmentSource;
+import org.cathori.backend.notice.infra.crawling.source.DepartmentSource;
 import org.cathori.backend.notice.model.Notice;
 import org.cathori.backend.notice.model.NoticeRepository;
 import org.cathori.backend.tag.domain.Tag;
